pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quant-edge-explorer"
version = "0.1.0"
description = "Conversational research agent for the quant-edge platform"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.88.0",
    "fantasy-data",
    "knowledge-base",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    "pandas>=2.1.0",
    "tabulate>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "scipy>=1.11.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.24.0", "ruff"]

[project.scripts]
explore = "explorer.main:main"
quant-edge = "server.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/explorer", "src/server"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
fantasy-data = { path = "../fantasy-data", editable = true }
knowledge-base = { path = "../knowledge-base", editable = true }
